refactor(splash_frame): split LoadingFrame into mixin package
Move widgets/frames/splash_frame.py into a package: frame.py builds the
UI, properties.py exposes set_ready, and handlers.py houses the timer
tick and _maybe_finish lifecycle transition.
