+++ widgets/frames/splash_frame/__init__.py
@@ -1,0 +1,7 @@
+"""Loading splash frame package."""
+
+from __future__ import annotations
+
+from widgets.frames.splash_frame.frame import LoadingFrame
+
+__all__ = ["LoadingFrame"]
